FIX error "ROUGE-1.5.5.pl returned non-zero exit status 255"

+++ tools/test_rouge.py
@@ -1,3 +1,4 @@
+# -*- encoding: utf-8 -*-
 import argparse
 import os
 import time
@@ -10,30 +11,38 @@ def test_rouge(cand_file, ref_file):
     f_ref = open(ref_file, encoding="utf-8")
     current_time = time.strftime('%Y-%m-%d-%H-%M-%S', time.localtime())
     tmp_dir = ".rouge-tmp-{}".format(current_time)
-    if not os.path.isdir(tmp_dir):
-        os.mkdir(tmp_dir)
-        os.mkdir(tmp_dir + "/candidate")
-        os.mkdir(tmp_dir + "/reference")
-    for i, line in enumerate(f_cand):
-        with open(tmp_dir + "/candidate/cand.{}.txt".format(
-                str(i).zfill(3)), "w", encoding="utf-8") as f:
-            f.write(line)
-    for i, line in enumerate(f_ref):
-        with open(tmp_dir + "/reference/ref.{}.txt".format(
-                str(i).zfill(3)), "w", encoding="utf-8") as f:
-            f.write(line)
-    f_cand.close()
-    f_ref.close()
-    r = pyrouge.Rouge155()
-    r.model_dir = tmp_dir + "/reference/"
-    r.system_dir = tmp_dir + "/candidate/"
-    r.model_filename_pattern = 'ref.#ID#.txt'
-    r.system_filename_pattern = 'cand.(\d+).txt'
-    rouge_results = r.convert_and_evaluate()
-    results_dict = r.output_to_dict(rouge_results)
-
-    shutil.rmtree(tmp_dir)
-    return results_dict
+    try:
+        if not os.path.isdir(tmp_dir):
+            os.mkdir(tmp_dir)
+            os.mkdir(tmp_dir + "/candidate")
+            os.mkdir(tmp_dir + "/reference")
+        candidates = [line.strip() for line in f_cand]
+        references = [line.strip() for line in f_ref]
+        assert len(candidates) == len(references)
+        cnt = len(candidates)
+        for i in range(cnt):
+            if len(references[i]) < 1:
+                continue
+            with open(tmp_dir + "/candidate/cand.{}.txt".format(i), "w",
+                      encoding="utf-8") as f:
+                f.write(candidates[i])
+            with open(tmp_dir + "/reference/ref.{}.txt".format(i), "w",
+                      encoding="utf-8") as f:
+                f.write(references[i])
+        f_cand.close()
+        f_ref.close()
+        r = pyrouge.Rouge155()
+        r.model_dir = tmp_dir + "/reference/"
+        r.system_dir = tmp_dir + "/candidate/"
+        r.model_filename_pattern = 'ref.#ID#.txt'
+        r.system_filename_pattern = 'cand.(\d+).txt'
+        rouge_results = r.convert_and_evaluate()
+        results_dict = r.output_to_dict(rouge_results)
+        return results_dict
+    finally:
+        pass
+        if os.path.isdir(tmp_dir):
+            shutil.rmtree(tmp_dir)
 
 
 if __name__ == "__main__":
